pyproject.toml
[tool.poetry]
name = "smartlock"
version = "0.1.0"
description = "ハードウェアから作るスマートロック"
authors = ["Takahiro55555 <tkhr.corei7@gmail.com>"]

[tool.poetry.dependencies]
python = "3.7"
"rpi.gpio" = "^0.7.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
